test: switch to `ReTestItems`
