+++ test/Project.toml
@@ -3,5 +3,6 @@ Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 HerbCore = "2b23ba43-8213-43cb-b5ea-38c12b45bd45"
 HerbGrammar = "4ef9e186-2fe5-4b24-8de7-9f7291f24af7"
 HerbSpecification = "6d54aada-062f-46d8-85cf-a1ceaf058a06"
+ReTestItems = "817f1d60-ba6b-4fd5-9520-3cf149f6a823"
 RuntimeGeneratedFunctions = "7e49a35a-f44a-4d26-94aa-eba1b4ca6b47"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
