+++ test/quality_tests.jl
@@ -1,0 +1,4 @@
+@testitem "Quality tests" begin
+    using Aqua
+    @testset "Aqua" Aqua.test_all(HerbInterpret)
+end
